add: use bun instead of nodejs
