+++ docker/dockerfiles/ocmstub.Dockerfile
@@ -1,4 +1,6 @@
-FROM node:lts-hydrogen
+# use the official Bun image
+# see all versions at https://hub.docker.com/r/oven/bun/tags
+FROM oven/bun:1-alpine
 
 # keys for oci taken from:
 # https://github.com/opencontainers/image-spec/blob/main/annotations.md#pre-defined-annotation-keys
@@ -7,32 +9,15 @@ LABEL org.opencontainers.image.title="PonderSource OCM Stub Image"
 LABEL org.opencontainers.image.source="https://github.com/pondersource/dev-stock"
 LABEL org.opencontainers.image.authors="Mohammad Mahdi Baghbani Pourvahid"
 
-# set timezone.
-ENV TZ=UTC
-RUN ln --symbolic --no-dereference --force /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
+WORKDIR /ocmstub
 
-ENV DEBIAN_FRONTEND noninteractive
-
-RUN apt update --yes
-
-# install dependencies.
-RUN apt install --yes               \
-    vim                             
-
-# CACHEBUST forces docker to clone fresh source codes from git.
-# example: docker build -t your-image --build-arg CACHEBUST="default" .
-# $RANDOM returns random number each time.
-ARG CACHEBUST="default"
-RUN git clone --depth 1 --branch main https://github.com/pondersource/ocm-stub
-
-WORKDIR /ocm-stub
+COPY ./scripts/ocmstub/index.js                                           index.js
 
 # trust all the certificates:
-COPY ./tls/certificates/*                                       /tls/
-COPY ./tls/certificate-authority/*                              /tls/
-RUN ln --symbolic --force /tls/*.crt                            /usr/local/share/ca-certificates
-RUN update-ca-certificates
+COPY ./tls/certificates/meshdir.crt                                       /tls/meshdir.crt
+COPY ./tls/certificates/meshdir.key                                       /tls/meshdir.key
 
-RUN npm install
-
-CMD NODE_TLS_REJECT_UNAUTHORIZED=0 node stub.js
+# run the app
+USER bun
+EXPOSE 443/tcp
+ENTRYPOINT [ "bun", "run", "index.js" ]
